Moving auth helpers into auth.js file and removing duplicate functions elsewhere.
