+++ resources/assets/selectors/user.js
@@ -1,24 +1,4 @@
 /**
- * Get the user id from the state.
- *
- * @param  {Object} state
- * @return {Bool}
- */
-export function isAuthenticated(state) {
-  return state.user.isAuthenticated;
-}
-
-/**
- * Get the user id from the state.
- *
- * @param  {Object} state
- * @return {String}
- */
-export function getUserId(state) {
-  return state.user.id;
-}
-
-/**
  * Get the user token from the state.
  *
  * @param  {Object} state
